\name{combineRTCA}
\Rdversion{1.1}
\alias{combineRTCA}
\title{
  Combine a list of RTCA objects
}
\description{
  Combine a list of RTCA objects
}
\usage{
combineRTCA(list)
}
\arguments{
  \item{list}{A list of \code{RTCA} objects}
}
\details{
  The current implementation requires all the objects have exactly the
  same time-points recorded (or at least of same length).
}
\value{
  A new \code{RTCA} object
}
\author{
  Jitao David Zhang \email{j.zhang@dkfz.de}
}
\examples{
}
